sign in/ out UI changes,a modal for artifact details, search box to filter loggers
