+++ monitoring-dashboard/components/org.wso2.micro.integrator.dashboard.web/web-app/src/commons/Modals/commons/LocalEntryValueSection.js
